+++ src/alfasim_sdk/_internal/alfacase/case_description_attributes.py
@@ -109,13 +109,13 @@ def to_scalar(
 
     Ex.:
     @attr.s
-    class TrendOutputDescription:
+    class PositionalOutputDescription:
         pos = attrib_scalar()
         temperature = attrib_scalar(is_optional=True))
 
-    TrendOutputDescription(position=Scalar(1,"m"))
-    TrendOutputDescription(position=(1,"m"))
-    TrendOutputDescription(temperature=None)
+    PositionalOutputDescription(position=Scalar(1,"m"))
+    PositionalOutputDescription(position=(1,"m"))
+    PositionalOutputDescription(temperature=None)
     """
     if is_optional and value is None:
         return value
@@ -249,6 +249,10 @@ def attrib_array(
 
         If a default is not set (``attr.NOTHING``), a value must be supplied when instantiating;
         otherwise, a `TypeError` will be raised.
+
+    :param category:
+        Name of the array category.
+
     """
     if isinstance(default, Array):
         if category is None:
@@ -279,6 +283,10 @@ def attrib_curve(
 
         If a default is not set (``attr.NOTHING``), a value must be supplied when instantiating;
         otherwise, a `TypeError` will be raised.
+
+    :param category:
+        Category of the curve image array.
+
     """
     if isinstance(default, Curve):
         if category is None: