+++ src/alfasim_sdk/_internal/constants.py
@@ -141,6 +141,7 @@ class NodeCellType(Enum):
     MassSource = "mass_source_boundary"
     Pressure = "pressure_boundary"
     Separator = "separator_node"
+    Controller = "controller_node"
 
 
 class MassInflowSplitType(Enum):
@@ -369,6 +370,10 @@ class WellConnectionPort(Enum):
     Top = "port"  # 'port' is being used for backwards compatibility.
     LeftAnnulus = "left_annulus_port"
     RightAnnulus = "right_annulus_port"
+
+
+class ControllerType(Enum):
+    PID = "pid"
 
 
 MULTI_INPUT_TYPE_SUFFIX = "_input_type"