+++ HISTORY.rst
@@ -2,6 +2,18 @@
 History
 =======
 
+
+0.11.0 (unreleased)
+===================
+
+* **Breaking Change**: Change ``TrendOutputDescription`` to support different trends types. Now trends in ``CaseOutputDescription`` are an object of ``TrendsOutputDescription`` that contains a list for each trend type. The available trend types are:
+    - ``PositionalPipeTrendDescription``;
+    - ``GlobalTrendDescription``;
+    - ``OverallPipeTrendDescription``;
+    - ``EquipmentTrendDescription``;
+    - ``SeparatorTrendDescription``;
+
+
 0.10.1 (2021-06-30)
 ===================
 
@@ -17,11 +29,11 @@ History
     - ``PressureNode`` and ``ReservoirInflowEquipment``: ``pressure``, ``temperature``, ``mass_fractions``, ``volume_fractions``, ``gas_liquid_ratio``, ``gas_oil_ratio``, ``water_cut``;
     - ``LinearIPR``: ``well_index``;
     - ``HeatSourceEquipment``: ``power``;
-* Add two new hooks to calculate solids model (for slurry viscosity and slip velocity)
+* Add two new hooks to calculate solids model (for slurry viscosity and slip velocity).
 * **Breaking Change**: Change ``OpeningCurveDescription`` (``opening_curve`` attribute) for ``Curve`` from barril.
 * **Breaking Change**: Change signature of ``HOOK_INITIALIZE_STATE_VARIABLES_CALCULATOR``.
 * **Breaking Change**: Change signature of ``HOOK_CALCULATE_RELATIVE_EMULSION_VISCOSITY``.
-* Add new API function ``get_relative_emulsion_viscosity`` which is a helper function that can be used in the Hooks of Liquid-Liquid Mechanistic Model
+* Add new API function ``get_relative_emulsion_viscosity`` which is a helper function that can be used in the Hooks of Liquid-Liquid Mechanistic Model.
 
 
 0.9.0 (2021-05-04)